requirements.txt
datasets==4.1.1
pandas==2.3.2
numpy==2.3.2
python-dotenv==1.1.1
tqdm==2.2.3